+++ cypress_shared/pages/apply/risk-strength-and-need-information/behaviour-in-cas/previousStays.ts
@@ -7,7 +7,7 @@ export default class PreviousStaysPage extends ApplyPage {
     super(
       'Behaviour in previous accommodation',
       application,
-      'behaviour-in-previous-accommodation',
+      'behaviour-in-cas',
       'previous-stays',
       paths.applications.show({ id: application.id }),
     )
