+++ server/form-pages/apply/accommodation-need/index.ts
@@ -1,5 +1,4 @@
 import { Section } from '../../utils/decorators'
-import BehaviourInPreviousAccommodation from './behaviour-in-previous-accommodation'
 import Consent from './consent'
 import ContactDetails from './contact-details'
 import Eligibility from './eligibility'
@@ -7,6 +6,6 @@ import SentenceInformation from './sentence-information'
 
 @Section({
   title: 'Accommodation need',
-  tasks: [SentenceInformation, ContactDetails, Eligibility, BehaviourInPreviousAccommodation, Consent],
+  tasks: [SentenceInformation, ContactDetails, Eligibility, Consent],
 })
 export default class AccommodationNeed {}
