+++ server/form-pages/apply/risk-strength-and-need-information/index.ts
@@ -1,11 +1,12 @@
 import { Section } from '../../utils/decorators'
 import ApprovalsForSpecificRisksTask from './approvals-for-specific-risks'
+import BehaviourInCas from './behaviour-in-cas'
 import LicenceConditions from './licence-conditions'
 import OasysImport from './oasys-import'
 import PrisonInformation from './prison-information'
 
 @Section({
   title: 'Risk, strength and need information',
-  tasks: [LicenceConditions, PrisonInformation, ApprovalsForSpecificRisksTask, OasysImport],
+  tasks: [LicenceConditions, PrisonInformation, ApprovalsForSpecificRisksTask, OasysImport, BehaviourInCas],
 })
 export default class RiskStrengthAndNeedInformation {}
